requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aspy.refactor-imports==2.2.0
    # via reorder-python-imports
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
attrs==21.4.0
    # via
    #   outcome
    #   pytest
    #   trio
black==21.12b0
    # via -r requirements.in
cached-property==1.5.2
    # via aspy.refactor-imports
certifi==2021.10.8
    # via urllib3
cffi==1.15.0
    # via cryptography
click==8.0.3
    # via black
cryptography==36.0.1
    # via
    #   pyopenssl
    #   urllib3
h11==0.12.0
    # via wsproto
idna==3.3
    # via
    #   trio
    #   urllib3
iniconfig==1.1.1
    # via pytest
mypy-extensions==0.4.3
    # via black
numpy==1.22.1
    # via opencv-python
obs-websocket-py==0.5.3
    # via -r requirements.in
opencv-python==4.5.5.62
    # via -r requirements.in
outcome==1.1.0
    # via trio
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
pillow==9.0.0
    # via -r requirements.in
platformdirs==2.4.1
    # via black
pluggy==1.0.0
    # via pytest
psutil==5.9.0
    # via -r requirements.in
py==1.11.0
    # via pytest
pycparser==2.21
    # via cffi
pydub==0.25.1
    # via -r requirements.in
pyopenssl==21.0.0
    # via urllib3
pyparsing==3.0.6
    # via packaging
pytest==6.2.5
    # via -r requirements.in
reorder-python-imports==2.6.0
    # via -r requirements.in
selenium==4.1.0
    # via -r requirements.in
six==1.16.0
    # via
    #   obs-websocket-py
    #   pyopenssl
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via trio
streamdeck==0.9.0
    # via -r requirements.in
toml==0.10.2
    # via pytest
tomli==1.2.3
    # via black
trio==0.19.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
typing-extensions==4.0.1
    # via black
urllib3[secure]==1.26.8
    # via selenium
websocket-client==1.2.3
    # via obs-websocket-py
wsproto==1.0.0
    # via trio-websocket
